backend/requirements.txt
fastapi
alpaca-trade-api
pandas
numpy
matplotlib
plotly
ta
python-dotenv
